Move responsibility classes into their own files
