+++ src/main/kotlin/org/vaccineimpact/api/models/responsibilities/ResponsibilitiesAndTouchstoneStatus.kt
@@ -1,0 +1,8 @@
+package org.vaccineimpact.api.models.responsibilities
+
+import org.vaccineimpact.api.models.TouchstoneStatus
+
+data class ResponsibilitiesAndTouchstoneStatus(
+        val responsibilities: Responsibilities,
+        val touchstoneStatus: TouchstoneStatus
+)
