+++ src/main/kotlin/org/vaccineimpact/api/models/responsibilities/ResponsibilityAndTouchstone.kt
@@ -1,0 +1,8 @@
+package org.vaccineimpact.api.models.responsibilities
+
+import org.vaccineimpact.api.models.TouchstoneVersion
+
+data class ResponsibilityAndTouchstone(
+        val touchstoneVersion: TouchstoneVersion,
+        val responsibility: Responsibility
+)
