+++ src/main/kotlin/org/vaccineimpact/api/models/responsibilities/ResponsibilitySet.kt
@@ -1,0 +1,8 @@
+package org.vaccineimpact.api.models.responsibilities
+
+data class ResponsibilitySet(
+        val modellingGroupId: String,
+        val touchstoneVersion: String,
+        val status: ResponsibilitySetStatus,
+        val responsibilities: List<Responsibility>
+)
