+++ src/main/kotlin/org/vaccineimpact/api/models/responsibilities/ResponsibilitySetStatus.kt
@@ -1,0 +1,9 @@
+package org.vaccineimpact.api.models.responsibilities
+
+enum class ResponsibilitySetStatus
+{
+    NOT_APPLICABLE,
+    INCOMPLETE,
+    SUBMITTED,
+    APPROVED
+}
