+++ src/main/kotlin/org/vaccineimpact/api/models/responsibilities/ResponsibilityStatus.kt
@@ -1,0 +1,8 @@
+package org.vaccineimpact.api.models.responsibilities
+
+enum class ResponsibilityStatus
+{
+    EMPTY,
+    INVALID,
+    VALID
+}
